move pkg/generate/app/cmd -> pkg/oc/genneratee/app/cmd
