+++ pkg/oc/generate/app/cmd/newapp_test.go
