+++ pkg/oc/generate/app/cmd/resolve.go
