+++ pkg/oc/generate/app/cmd/resolve_test.go
